requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astrapy==1.4.1
attrs==24.2.0
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
frozenlist==1.4.1
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-astradb==0.3.1
langchain-core==0.2.34
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langchainhub==0.1.21
langsmith==0.1.104
multidict==6.0.5
numpy==1.26.4
openai==1.42.0
orjson==3.10.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.5
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
uuid6==2024.1.12
yarl==1.9.4
langchain_community==0.2.12
